refactor: update supabase client import paths for consistency
The import paths for the supabase client were updated to use a consistent path structure across the application. This change improves maintainability and reduces confusion by standardizing the import paths.
