+++ project/app/protected/library/manage-items/update-item/page.tsx
@@ -1,4 +1,4 @@
-import { createClient } from '@/utils/supabase/server'
+import { createClient } from '@/supabase/server'
 
 export default async function Page() {
   const supabase = await createClient()
