+++ project/app/protected/library/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { createClient } from "@/utils/supabase/client";
+import { createClient } from "@/supabase/client";
 import { useEffect, useState } from "react";
 import Link from 'next/link';
